Separated VisualizationHelpers.ts into types files
