+++ src/modules/visualization/visualizations/chart/Chart.vue
@@ -6,12 +6,10 @@ import ChartJsView from 'osh-js/source/core/ui/view/chart/ChartJsView.js';
 import CurveLayer from 'osh-js/source/core/ui/layer/CurveLayer.js';
 import { OSHVisualization } from '@/lib/OSHConnectDataStructs';
 import { createDatasource } from '@/modules/visualization/services/datasource.service';
-import {
-	IChartViewProperties,
-	ICurveLayerProperties,
-	ISweApiDataSourceProperties,
-} from '@/lib/VisualizationHelpers';
 import { useVisualizationCleanup } from '../../sidebar/composables/useVisualizationCleanup';
+import { ISweApiDataSourceProperties } from '../../types/datasource';
+import { ICurveLayerProperties } from '../../types/layers';
+import { IChartViewProperties } from '../../types/views';
 
 const props = defineProps<{
 	visualization: OSHVisualization;
