+++ src/modules/visualization/visualizations/text/Builder.ts
@@ -1,10 +1,5 @@
 import { useVisualizationStore } from '@/stores/visualizationstore';
 import { useVizWizStore } from '@/stores/vizwizstore';
-import {
-	DataLayerProperties,
-	ISweApiDataSourceProperties,
-	VisualizationComponents,
-} from '@/lib/VisualizationHelpers';
 import { OSHVisualization } from '@/lib/OSHConnectDataStructs';
 //@ts-ignore
 import { Mode } from 'osh-js/source/core/datasource/Mode';
@@ -17,6 +12,9 @@ import {
 	BuildRoleProperty,
 	getUsedDatastreams,
 } from '../../services/aggregation.service';
+import { VisualizationComponents } from '../../types/visualization';
+import { ISweApiDataSourceProperties } from '../../types/datasource';
+import { DataLayerProperties } from '../../types/layers';
 
 export default function build() {
 	console.log('Building Text Visualization...');
@@ -32,7 +30,6 @@ export default function build() {
 	const visualizationComponents: VisualizationComponents = {
 		dataSource: textResult.vizDatasources,
 		dataLayer: textResult.dataLayer,
-		dataView: null,
 	};
 
 	const newViz = new OSHVisualization(
