+++ src/modules/visualization/visualizations/text/Text.vue
@@ -4,8 +4,8 @@ import { onMounted, ref } from 'vue';
 import { randomUUID } from 'osh-js/source/core/utils/Utils.js';
 import { DATASOURCE_DATA_TOPIC } from 'osh-js/source/core/Constants.js';
 import { createDatasource } from '@/modules/visualization/services/datasource.service';
-import { ISweApiDataSourceProperties } from '@/lib/VisualizationHelpers';
 import { useVisualizationCleanup } from '../../sidebar/composables/useVisualizationCleanup';
+import { ISweApiDataSourceProperties } from '../../types/datasource';
 
 // Generate a random ID when the component is created
 const textboxId = ref('textbox-' + randomUUID());
